copy_splice(): don't leak pipe fds if splice() fails
Previously, if either splice() call failed, the function would return
without closing pin or pout.  (One way it might fail is if the process
reading from the file system closes its fd without consuming the whole
file.)

To fix this, wrap the pair of fds in a new struct PipeFds with an
idempotent close() function, () and a Drop impl that calls it.  This
allows copy_splice() to propagate errors encountered while closing
either pipe fd after an otherwise-successful copy, while guaranteeing
that close() will be called on both pipe fds.

If close() fails on both pipe fds, return the error from the 'in' side,
which matches the old behaviour.

This addresses the fd-leak part of #18.